+++ react-voice-commons-sdk/lib/index.d.ts
@@ -11,6 +11,7 @@
 export {
   TelnyxVoipClient,
   createTelnyxVoipClient,
+  destroyTelnyxVoipClient,
   createBackgroundTelnyxVoipClient,
 } from './telnyx-voip-client';
 export type { TelnyxVoipClientOptions } from './telnyx-voip-client';